feat(db): drop *Threshold4 columns — migration destructive #3352 (#3364)
