+++ packages/app/drizzle/meta/_journal.json
@@ -204,6 +204,13 @@
 			"when": 1776000000000,
 			"tag": "0028_add_cse_opinion_completed_at",
 			"breakpoints": true
+		},
+		{
+			"idx": 29,
+			"version": "7",
+			"when": 1777494411210,
+			"tag": "0029_drop_quartile_threshold4",
+			"breakpoints": true
 		}
 	]
 }
