+++ packages/app/src/app/declaration-remuneration/etape/[step]/page.tsx
@@ -86,7 +86,7 @@ export default async function StepPage({ params }: StepPageProps) {
 				men: d.indicatorFAnnualMen3 ?? undefined,
 			},
 			{
-				threshold: d.indicatorFAnnualThreshold4 ?? "",
+				threshold: "",
 				women: d.indicatorFAnnualWomen4 ?? undefined,
 				men: d.indicatorFAnnualMen4 ?? undefined,
 			},
@@ -108,7 +108,7 @@ export default async function StepPage({ params }: StepPageProps) {
 				men: d.indicatorFHourlyMen3 ?? undefined,
 			},
 			{
-				threshold: d.indicatorFHourlyThreshold4 ?? "",
+				threshold: "",
 				women: d.indicatorFHourlyWomen4 ?? undefined,
 				men: d.indicatorFHourlyMen4 ?? undefined,
 			},
