+++ packages/app/src/modules/declaration-remuneration/shared/gipMdsMapping.ts
@@ -8,8 +8,8 @@ export type GipMdsRow = typeof gipMdsData.$inferSelect;
 
 /** Quartile data computed from GIP proportions + workforce totals. */
 export type GipQuartileData = {
-	/** 4 thresholds, one per quartile (lower bound of each quartile group). */
-	thresholds: [string | null, string | null, string | null, string | null];
+	/** 3 thresholds for Q1-Q3 (lower bound); Q4 has no threshold in the GIP model. */
+	thresholds: [string | null, string | null, string | null];
 	/** Integer women count per quartile, derived from proportion × total/4. */
 	womenCounts: [number | null, number | null, number | null, number | null];
 	/** Integer men count per quartile, derived from proportion × total/4. */
@@ -109,7 +109,6 @@ export function mapGipToFormData(row: GipMdsRow | null): GipPrefillData | null {
 					row.annualQuartileThreshold1,
 					row.annualQuartileThreshold2,
 					row.annualQuartileThreshold3,
-					row.annualQuartileThreshold4,
 				],
 				[
 					row.annualQuartile1ProportionWomen,
@@ -131,7 +130,6 @@ export function mapGipToFormData(row: GipMdsRow | null): GipPrefillData | null {
 					row.hourlyQuartileThreshold1,
 					row.hourlyQuartileThreshold2,
 					row.hourlyQuartileThreshold3,
-					row.hourlyQuartileThreshold4,
 				],
 				[
 					row.hourlyQuartile1ProportionWomen,
@@ -194,7 +192,6 @@ export const CSV_TO_SCHEMA_MAP: Record<string, keyof GipMdsRow> = {
 	Seuil_Q1_Rem_globale: "annualQuartileThreshold1",
 	Seuil_Q2_Rem_globale: "annualQuartileThreshold2",
 	Seuil_Q3_Rem_globale: "annualQuartileThreshold3",
-	Seuil_Q4_Rem_globale: "annualQuartileThreshold4",
 	Quartile1_Rem_globale_annuelle_proportion_F: "annualQuartile1ProportionWomen",
 	Quartile2_Rem_globale_annuelle_proportion_F: "annualQuartile2ProportionWomen",
 	Quartile3_Rem_globale_annuelle_proportion_F: "annualQuartile3ProportionWomen",
@@ -206,7 +203,6 @@ export const CSV_TO_SCHEMA_MAP: Record<string, keyof GipMdsRow> = {
 	Seuil_Q1_Taux_horaire_global: "hourlyQuartileThreshold1",
 	Seuil_Q2_Taux_horaire_global: "hourlyQuartileThreshold2",
 	Seuil_Q3_Taux_horaire_global: "hourlyQuartileThreshold3",
-	Seuil_Q4_Taux_horaire_global: "hourlyQuartileThreshold4",
 	Quartile1_Taux_horaire_global_proportion_F: "hourlyQuartile1ProportionWomen",
 	Quartile2_Taux_horaire_global_proportion_F: "hourlyQuartile2ProportionWomen",
 	Quartile3_Taux_horaire_global_proportion_F: "hourlyQuartile3ProportionWomen",
@@ -261,7 +257,7 @@ function proportionToCount(
 function buildQuartileData(
 	totalWomen: number | null,
 	totalMen: number | null,
-	thresholds: [string | null, string | null, string | null, string | null],
+	thresholds: [string | null, string | null, string | null],
 	womenProportions: [
 		string | null,
 		string | null,
