+++ packages/app/src/modules/declaration-remuneration/steps/__tests__/Step2PayGap.test.tsx
@@ -196,12 +196,12 @@ describe("Step2PayGap", () => {
 					},
 					step4: {
 						annual: {
-							thresholds: [null, null, null, null],
+							thresholds: [null, null, null],
 							womenCounts: [null, null, null, null],
 							menCounts: [null, null, null, null],
 						},
 						hourly: {
-							thresholds: [null, null, null, null],
+							thresholds: [null, null, null],
 							womenCounts: [null, null, null, null],
 							menCounts: [null, null, null, null],
 						},
