+++ packages/app/src/modules/export/fetchDeclarations.ts
@@ -99,13 +99,13 @@ export function buildIndicators(row: DeclarationRow) {
 		row.indicatorFAnnualThreshold1,
 		row.indicatorFAnnualThreshold2,
 		row.indicatorFAnnualThreshold3,
-		row.indicatorFAnnualThreshold4,
+		null,
 	];
 	const hourlyThresholds = [
 		row.indicatorFHourlyThreshold1,
 		row.indicatorFHourlyThreshold2,
 		row.indicatorFHourlyThreshold3,
-		row.indicatorFHourlyThreshold4,
+		null,
 	];
 
 	const annualQuartile = Object.fromEntries(
