+++ packages/app/src/modules/export/queries.ts
@@ -39,7 +39,6 @@ export const indicatorColumns = {
 	indicatorFAnnualThreshold1: declarations.indicatorFAnnualThreshold1,
 	indicatorFAnnualThreshold2: declarations.indicatorFAnnualThreshold2,
 	indicatorFAnnualThreshold3: declarations.indicatorFAnnualThreshold3,
-	indicatorFAnnualThreshold4: declarations.indicatorFAnnualThreshold4,
 	indicatorFAnnualWomen1: declarations.indicatorFAnnualWomen1,
 	indicatorFAnnualWomen2: declarations.indicatorFAnnualWomen2,
 	indicatorFAnnualWomen3: declarations.indicatorFAnnualWomen3,
@@ -51,7 +50,6 @@ export const indicatorColumns = {
 	indicatorFHourlyThreshold1: declarations.indicatorFHourlyThreshold1,
 	indicatorFHourlyThreshold2: declarations.indicatorFHourlyThreshold2,
 	indicatorFHourlyThreshold3: declarations.indicatorFHourlyThreshold3,
-	indicatorFHourlyThreshold4: declarations.indicatorFHourlyThreshold4,
 	indicatorFHourlyWomen1: declarations.indicatorFHourlyWomen1,
 	indicatorFHourlyWomen2: declarations.indicatorFHourlyWomen2,
 	indicatorFHourlyWomen3: declarations.indicatorFHourlyWomen3,
