+++ packages/app/src/server/api/routers/declaration.ts
@@ -198,7 +198,6 @@ export const declarationRouter = createTRPCRouter({
 									indicatorFAnnualThreshold1: null,
 									indicatorFAnnualThreshold2: null,
 									indicatorFAnnualThreshold3: null,
-									indicatorFAnnualThreshold4: null,
 									indicatorFAnnualWomen1: null,
 									indicatorFAnnualWomen2: null,
 									indicatorFAnnualWomen3: null,
@@ -210,7 +209,6 @@ export const declarationRouter = createTRPCRouter({
 									indicatorFHourlyThreshold1: null,
 									indicatorFHourlyThreshold2: null,
 									indicatorFHourlyThreshold3: null,
-									indicatorFHourlyThreshold4: null,
 									indicatorFHourlyWomen1: null,
 									indicatorFHourlyWomen2: null,
 									indicatorFHourlyWomen3: null,
@@ -298,7 +296,6 @@ export const declarationRouter = createTRPCRouter({
 					indicatorFAnnualThreshold1: input.annual[0].threshold || null,
 					indicatorFAnnualThreshold2: input.annual[1].threshold || null,
 					indicatorFAnnualThreshold3: input.annual[2].threshold || null,
-					indicatorFAnnualThreshold4: input.annual[3].threshold || null,
 					indicatorFAnnualWomen1: input.annual[0].women ?? null,
 					indicatorFAnnualWomen2: input.annual[1].women ?? null,
 					indicatorFAnnualWomen3: input.annual[2].women ?? null,
@@ -310,7 +307,6 @@ export const declarationRouter = createTRPCRouter({
 					indicatorFHourlyThreshold1: input.hourly[0].threshold || null,
 					indicatorFHourlyThreshold2: input.hourly[1].threshold || null,
 					indicatorFHourlyThreshold3: input.hourly[2].threshold || null,
-					indicatorFHourlyThreshold4: input.hourly[3].threshold || null,
 					indicatorFHourlyWomen1: input.hourly[0].women ?? null,
 					indicatorFHourlyWomen2: input.hourly[1].women ?? null,
 					indicatorFHourlyWomen3: input.hourly[2].women ?? null,
