+++ packages/app/src/server/db/schema.ts
@@ -85,7 +85,6 @@ export const declarations = createTable(
 		indicatorFAnnualThreshold1: d.numeric(),
 		indicatorFAnnualThreshold2: d.numeric(),
 		indicatorFAnnualThreshold3: d.numeric(),
-		indicatorFAnnualThreshold4: d.numeric(),
 		indicatorFAnnualWomen1: d.integer(),
 		indicatorFAnnualWomen2: d.integer(),
 		indicatorFAnnualWomen3: d.integer(),
@@ -98,7 +97,6 @@ export const declarations = createTable(
 		indicatorFHourlyThreshold1: d.numeric(),
 		indicatorFHourlyThreshold2: d.numeric(),
 		indicatorFHourlyThreshold3: d.numeric(),
-		indicatorFHourlyThreshold4: d.numeric(),
 		indicatorFHourlyWomen1: d.integer(),
 		indicatorFHourlyWomen2: d.integer(),
 		indicatorFHourlyWomen3: d.integer(),
@@ -287,7 +285,6 @@ export const gipMdsData = createTable(
 		annualQuartileThreshold1: d.numeric({ precision: 9, scale: 2 }),
 		annualQuartileThreshold2: d.numeric({ precision: 9, scale: 2 }),
 		annualQuartileThreshold3: d.numeric({ precision: 9, scale: 2 }),
-		annualQuartileThreshold4: d.numeric({ precision: 9, scale: 2 }),
 		annualQuartile1ProportionWomen: d.numeric({ precision: 9, scale: 4 }),
 		annualQuartile2ProportionWomen: d.numeric({ precision: 9, scale: 4 }),
 		annualQuartile3ProportionWomen: d.numeric({ precision: 9, scale: 4 }),
@@ -300,7 +297,6 @@ export const gipMdsData = createTable(
 		hourlyQuartileThreshold1: d.numeric({ precision: 9, scale: 2 }),
 		hourlyQuartileThreshold2: d.numeric({ precision: 9, scale: 2 }),
 		hourlyQuartileThreshold3: d.numeric({ precision: 9, scale: 2 }),
-		hourlyQuartileThreshold4: d.numeric({ precision: 9, scale: 2 }),
 		hourlyQuartile1ProportionWomen: d.numeric({ precision: 9, scale: 4 }),
 		hourlyQuartile2ProportionWomen: d.numeric({ precision: 9, scale: 4 }),
 		hourlyQuartile3ProportionWomen: d.numeric({ precision: 9, scale: 4 }),
